+++ js/modal.js
@@ -1,6 +1,3 @@
-import { posts } from "./data.js";
-import { renderPosts } from "./preview.js";
-
 const MODAL_SHOW_TIME = 3000;
 
 const receiveErrorTemplate = document.querySelector('#error-receive').content.querySelector('.error');
