+++ js/validation.js
@@ -1,4 +1,5 @@
-import { POSTS_DATA } from "./data.js";
+import { POSTS_DATA } from './data.js';
+import { getWordEnding } from './util.js';
 
 const Callbacks = {
   'hashtag': (text) => validationHashtag(text),
@@ -60,10 +61,10 @@ const validationHashtag = (text) => {
 
 const validationDescription = (text) => {
   const textLength = text.trim().length;
-  const maxLength = POSTS_DATA.commentMaxLength;
+  const maxLength = POSTS_DATA.comment_max_length;
 
   if (textLength > maxLength) {
-    return `Превышен лимит символов: ${maxLength}. Удлите `;
+    return `Превышен лимит символов: ${maxLength}. Удалите ${(textLength - maxLength)} ${getWordEnding(textLength - maxLength, ['символ', 'символа', 'символов'])}.`;
   }
 };
